Добавить интерфейсы для всех сервисов,с реализацией в отдельных классах.
