+++ src/main/java/ru/yandex/practicum/filmorate/service/IdGenerator.java
@@ -1,5 +1,10 @@
 package ru.yandex.practicum.filmorate.service;
 
+import org.springframework.context.annotation.Scope;
+import org.springframework.stereotype.Component;
+
+@Component
+@Scope("prototype")
 public class IdGenerator {
 
     private Integer id;
